Changed the location of the `conform-zod.spec.ts` file
