+++ packages/conform-zod/.gitignore
@@ -1,7 +1,2 @@
-_virtual
-*.cjs
-*.d.ts
-*.js
-*.mjs
+dist
 README
-!rollup.config.js
