+++ packages/conform-zod/rollup.config.js
@@ -6,7 +6,7 @@ import copy from 'rollup-plugin-copy';
 /** @returns {import("rollup").RollupOptions[]} */
 function configurePackage() {
 	let sourceDir = '.';
-	let outputDir = sourceDir;
+	let outputDir = './dist';
 
 	/** @type {import("rollup").RollupOptions} */
 	let ESM = {
