+++ packages/conform-zod/tests/helpers.ts
@@ -1,0 +1,11 @@
+export function createFormData(
+	entries: Array<[string, FormDataEntryValue]>,
+): FormData {
+	const formData = new FormData();
+
+	for (const [name, value] of entries) {
+		formData.append(name, value);
+	}
+
+	return formData;
+}
