+++ packages/conform-zod/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+	"extends": "./tsconfig",
+	"compilerOptions": {
+		"outDir": "./dist"
+	},
+	"exclude": ["./tests/**", "dist"]
+}
